Minor changes and improvements
