+++ src/main/java/com/minecraftmarket/minecraftmarket/bungee/MCMarket.java
@@ -3,10 +3,10 @@ package com.minecraftmarket.minecraftmarket.bungee;
 import com.minecraftmarket.minecraftmarket.bungee.commands.MMCmd;
 import com.minecraftmarket.minecraftmarket.bungee.configs.MainConfig;
 import com.minecraftmarket.minecraftmarket.bungee.tasks.PurchasesTask;
-import com.minecraftmarket.minecraftmarket.bungee.utils.updater.Updater;
 import com.minecraftmarket.minecraftmarket.common.api.MCMarketApi;
 import com.minecraftmarket.minecraftmarket.common.i18n.I18n;
 import com.minecraftmarket.minecraftmarket.common.metrics.BungeeMetrics;
+import com.minecraftmarket.minecraftmarket.common.updater.UpdateChecker;
 import com.minecraftmarket.minecraftmarket.common.utils.FileUtils;
 import net.md_5.bungee.api.plugin.Plugin;
 
@@ -32,10 +32,10 @@ public final class MCMarket extends Plugin {
         getProxy().getPluginManager().registerCommand(this, new MMCmd(this));
 
         new BungeeMetrics(this);
-        new Updater(this, 44031, pluginURL -> {
+        getProxy().getScheduler().runAsync(this, () -> new UpdateChecker(getDescription().getVersion(), 44031, pluginURL -> {
             getLogger().warning(I18n.tl("new_version"));
             getLogger().warning(pluginURL);
-        });
+        }));
     }
 
     @Override
